Improve dashboard layout when only one or two values exist
With a single visible metric the fixed two-up grid stranded a card in a
half-empty row and left most of the screen blank. The dashboard now:

- collapses to a single full-width column when there are 1-2 metrics, so
  cards read as intentional larger cards instead of a sparse grid
- shows a gentle 'Track Your Trends' hint card (reusing existing strings)
  that explains why no sparklines appear yet and offers a one-tap path to
  import another report

Extracted MetricCard and its supporting types into MetricCard.swift to
keep DashboardView under the 500-line file-length limit.
