+++ LabImporter/Views/MetricCard.swift
@@ -1,0 +1,164 @@
+import SwiftUI
+import Charts
+
+// MARK: - Supporting types
+
+struct SparkPoint: Identifiable {
+    let id = UUID()
+    let date: Date
+    let value: Double
+}
+
+struct MetricData: Identifiable {
+    var id: String { entry.code }
+    let entry: LabReport.Entry
+    let history: [SparkPoint]
+}
+
+// MARK: - MetricCard
+
+struct MetricCard: View {
+    let metric: MetricData
+    let isPinned: Bool
+
+    private var categoryColor: Color {
+        LabCategory.forCode(metric.entry.code).color
+    }
+
+    /// Direction of the latest reading relative to the previous one, used to show
+    /// a small trend arrow in the overview. `nil` when there is no prior value to
+    /// compare against.
+    private enum Trend {
+        case rising, falling, steady
+
+        var symbol: String {
+            switch self {
+            case .rising: return "arrow.up.right"
+            case .falling: return "arrow.down.right"
+            case .steady: return "arrow.right"
+            }
+        }
+
+        var accessibilityLabel: Text {
+            switch self {
+            case .rising: return Text("Trending up")
+            case .falling: return Text("Trending down")
+            case .steady: return Text("No change")
+            }
+        }
+    }
+
+    private var trend: Trend? {
+        guard metric.history.count > 1 else { return nil }
+        let latest = metric.history[metric.history.count - 1].value
+        let previous = metric.history[metric.history.count - 2].value
+        if latest > previous { return .rising }
+        if latest < previous { return .falling }
+        return .steady
+    }
+
+    /// The colored category "dock" at the card's top-left. When a trend is
+    /// available it doubles as the trend indicator and hosts a directional
+    /// arrow; otherwise it is a simple category-color dot filling the same
+    /// circle. Both variants share an 18×18 footprint so the title alignment —
+    /// and the dot/arrow size — stays consistent across cards.
+    @ViewBuilder
+    private var dock: some View {
+        if let trend {
+            Image(systemName: trend.symbol)
+                .font(.system(size: 9, weight: .heavy))
+                .foregroundStyle(.white)
+                .frame(width: 18, height: 18)
+                .background(categoryColor.gradient, in: Circle())
+                .accessibilityLabel(trend.accessibilityLabel)
+        } else {
+            Circle()
+                .fill(categoryColor.gradient)
+                .frame(width: 18, height: 18)
+                .accessibilityHidden(true)
+        }
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 6) {
+            HStack(alignment: .top, spacing: 7) {
+                dock
+                Text(metric.entry.resolvedName)
+                    .font(.caption)
+                    .foregroundStyle(.secondary)
+                    .lineLimit(2)
+                    .fixedSize(horizontal: false, vertical: true)
+                    .padding(.top, 1)
+                Spacer(minLength: 0)
+                if isPinned {
+                    Image(systemName: "pin.fill")
+                        .font(.caption2)
+                        .foregroundStyle(Color.yellow)
+                }
+            }
+
+            HStack(alignment: .firstTextBaseline, spacing: 3) {
+                Text(metric.entry.displayValue)
+                    .font(.title2.bold())
+                    .foregroundStyle(.primary)
+                    .lineLimit(1)
+                    .minimumScaleFactor(0.7)
+                if !metric.entry.unit.isEmpty {
+                    Text(metric.entry.unit)
+                        .font(.caption)
+                        .foregroundStyle(.secondary)
+                        .lineLimit(1)
+                }
+                Spacer(minLength: 0)
+            }
+
+            if metric.history.count > 1 {
+                sparkline
+            } else {
+                Spacer(minLength: 44)
+            }
+        }
+        .padding(14)
+        .frame(maxWidth: .infinity, minHeight: 158, alignment: .topLeading)
+        .background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 20))
+        .overlay(
+            RoundedRectangle(cornerRadius: 20)
+                .stroke(Color.primary.opacity(0.1), lineWidth: 0.5)
+        )
+    }
+
+    private var sparkline: some View {
+        Chart(metric.history) { point in
+            LineMark(
+                x: .value("Date", point.date),
+                y: .value("Value", point.value)
+            )
+            .foregroundStyle(categoryColor.opacity(0.85))
+
+            AreaMark(
+                x: .value("Date", point.date),
+                y: .value("Value", point.value)
+            )
+            .foregroundStyle(
+                LinearGradient(
+                    colors: [categoryColor.opacity(0.3), categoryColor.opacity(0.05)],
+                    startPoint: .top,
+                    endPoint: .bottom
+                )
+            )
+
+            PointMark(
+                x: .value("Date", point.date),
+                y: .value("Value", point.value)
+            )
+            .foregroundStyle(categoryColor)
+            .symbolSize(20)
+        }
+        .chartXAxis(.hidden)
+        .chartYAxis(.hidden)
+        // Grow to absorb the card's remaining vertical space instead of leaving a
+        // fixed-height chart with blank padding beneath it. `minHeight` keeps a
+        // sensible floor when a card's row is short.
+        .frame(minHeight: 44, maxHeight: .infinity)
+    }
+}
